% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize_with_sign.R
\name{binarize_with_sign}
\alias{binarize_with_sign}
\title{binarize_with_sign}
\usage{
binarize_with_sign(data, threshold)
}
\arguments{
\item{data}{numerical vector}

\item{threshold}{threshold used to binarise the data}
}
\description{
binarizes the data based on the threshold using the absolute value of the data.
Gives a sign based on the sign of the input data.
}
