% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_CARNIVAL_options.R
\name{check_CARNIVAL_options}
\alias{check_CARNIVAL_options}
\title{Check CARNIVAL Options}
\usage{
check_CARNIVAL_options(opts)
}
\arguments{
\item{opts}{An object of type \dQuote{\code{list}} defining the run parameters
for CARNIVAL.  Use the \code{\link{default_CARNIVAL_options}} function to 
create a list with default parameter settings. If cplex or cbc are chosen 
as the solver, the parameter solverPath needs to be supplied (not 
automatically added by \code{default_CARNIVAL_options()}).}
}
\description{
Checks the list of input options for CARNIVAL for completeness.
}
\seealso{
\code{\link{default_CARNIVAL_options}},  
  \code{\link[CARNIVAL]{runCARNIVAL}}
}
