% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gene_names.R
\name{check_gene_names}
\alias{check_gene_names}
\title{check_gene_names}
\usage{
check_gene_names(...)
}
\arguments{
\item{...}{takes named vectors}
}
\description{
checks if gene names looks like an EntrezID with a prefix of X.
}
