% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs_CARNIVAL.R
\name{check_inputs_for_CARNIVAL}
\alias{check_inputs_for_CARNIVAL}
\title{Check Inputs For CARNIVAL}
\usage{
check_inputs_for_CARNIVAL(meta_network, input_data, measured_data)
}
\arguments{
\item{meta_network}{Prior knowledge network (PKN).  \dQuote{\code{data.frame}} 
object with source, sign and target columns.  By default COSMOS uses a PKN 
derived from Omnipath, STITCHdb and Recon3D.  See details on the data 
\code{\link{meta_network}}.}

\item{input_data}{Numerical vector, where names are input nodes in the PKN 
and values are from \{1, 0, -1\}.}

\item{measured_data}{Numerical vector, where names are measured nodes in the
PKN and values are continuous values.  These values are compared to with
the simulation.}
}
\description{
Checks the format of the main data inputs for CARNIVL.  Checks the data format
and coverage of nodes in the PKN and data.  Ensures that all nodes in 
input_data and measured_data apper in the PKN.
}
\seealso{
\code{\link{default_CARNIVAL_options}}
}
