% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_network_data_coverage.R
\name{check_network_data_coverage}
\alias{check_network_data_coverage}
\title{checks coverage between data and prior knowledge}
\usage{
check_network_data_coverage(
  meta_network,
  tf_regulon = NULL,
  signaling_data,
  metabolic_data,
  expression_data = NULL,
  verbose = TRUE
)
}
\arguments{
\item{meta_network}{contains the PKN}

\item{tf_regulon}{(optional) tf-target network with EntrezID}

\item{signaling_data}{numerical vector, where names are signaling nodes 
in the PKN and values are from \{1,-1\}}

\item{metabolic_data}{numerical vector, where names are metabolic nodes 
in the PKN and values are continuous values. These values are compared to 
with the the simulation [? range of values]}

\item{expression_data}{(optional) numerical vector, where names are gene names
and values are from \{1,-1\}}

\item{verbose}{(default: TRUE) reports coverage}
}
\description{
checks the following
- signaling data nodes are in PKN
- metabolic data nodes are in PKN
- expression data genes overlap with TF targets
Stops if a check fails
}
