% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructQuery.R
\name{constructQuery}
\alias{constructQuery}
\title{Set up initial parameters for Cosmos DB querying}
\usage{
constructQuery(sql.what, sql.where)
}
\arguments{
\item{sql.what}{String for specifying what fields to retrieve. Typically called select condition}

\item{sql.where}{String for specifying what filter to use on data. Typically called search condition}
}
\value{
String containing a full SELECT statement for the database
}
\description{
Set up initial parameters for Cosmos DB querying
}
\examples{
constructQuery(sql.what = "c.contact.eloquaId", sql.where = "c.contact.eloquaId != null")
}
\keyword{query}
\keyword{select}
\keyword{where}
