% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ensembl_to_entrezid.R
\name{convert_ensembl_to_entrezid}
\alias{convert_ensembl_to_entrezid}
\title{convert gene ensembl to entrez id}
\usage{
convert_ensembl_to_entrezid(ensembl)
}
\arguments{
\item{ensembl}{vector of genes with ensembl id}
}
\value{
named vector, where names are the old ensemblIDs and values are the 
entrezIDs
}
\description{
convert gene ensembl to entrez id
}
\examples{
ensembl <- c("ENSG00000100601", "ENSG00000178826", "ENSG00000138231")
entrez_map <- convert_ensembl_to_entrezid(ensembl)
}
\seealso{
\code{\link{convert_genesymbols_to_entrezid}}
}
