% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genesymbols_to_entrezid.R
\name{convert_genesymbols_to_entrezid}
\alias{convert_genesymbols_to_entrezid}
\title{convert gene symbols to entrez id}
\usage{
convert_genesymbols_to_entrezid(symbols)
}
\arguments{
\item{symbols}{vector of genesymbols}
}
\value{
data.frame with human gene ENTREZID and SYMBOL mapping
}
\description{
convert gene symbols to entrez id
}
\examples{
symbols <- c("MDH1", "PARP1", "IL6")
symbol_entrez_map <- convert_genesymbols_to_entrezid(symbols)
}
\seealso{
\code{\link{convert_ensembl_to_entrezid}}
}
