% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmosAuth.R
\name{cosmosAuth}
\alias{cosmosAuth}
\title{Set up initial parameters for Cosmos DB querying}
\usage{
cosmosAuth(key, uri, dbName, collName)
}
\arguments{
\item{key}{Access key, primary or secondary, for interacting with Cosmos DB collection}

\item{uri}{HTTPS URI to the Cosmos DB}

\item{dbName}{String name of CosmosDB Database target}

\item{collName}{String name of CosmosDB Collection target}
}
\value{
Environment variable is created for usage in all other functions
}
\description{
Set up initial parameters for Cosmos DB querying
}
\examples{
cosmosAuth("KeY123","https://yourcosmosdb.documents.azure.com","SampleDB","SampleCollection")
}
\keyword{cosmos}
\keyword{environment}
\keyword{key}
\keyword{uri}
