% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmosCreate.R
\name{cosmosCreate}
\alias{cosmosCreate}
\title{POST a Document Create to the REST API for Cosmos DB.}
\usage{
cosmosCreate(
  sql.doc = "",
  sql.partitionkey_value = "",
  debug.auth = FALSE,
  debug.query = FALSE,
  content.response = FALSE
)
}
\arguments{
\item{sql.doc}{List of key/values to add to document. Required.}

\item{sql.partitionkey_value}{The value of the partitionkey. Must also be part of sql.doc. Required.}

\item{debug.auth}{Logical value for getting verbose output of auth header being constructed. Defaults to false.}

\item{debug.query}{Logical value for getting verbose output of HTTP response, printing all headers. Defaults to false.}

\item{content.response}{Logical value to determine whether to retrieve full response or just the documents}
}
\value{
Prints status code of HTTP POST, and returns full HTTP response or just the content
}
\description{
POST a Document Create to the REST API for Cosmos DB.
}
\examples{
cosmosCreate(sql.doc = list(id="uuid", code="code1"), sql.partitionkey_value = "code1")
}
\keyword{cosmosdb}
\keyword{create}
\keyword{post}
