% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmosQuery.R
\name{cosmosQuery}
\alias{cosmosQuery}
\title{POST a full query to the REST API for Cosmos DB.}
\usage{
cosmosQuery(sql.what = "*", sql.where = "", sql.params = list(),
  max.items = 100, debug.auth = FALSE, debug.query = FALSE,
  content.response = FALSE, flatten = FALSE, max.retry = 5,
  retry.pause = 0)
}
\arguments{
\item{sql.what}{String for specifying what fields to retrieve. Typically called select condition. Defaults to *}

\item{sql.where}{String for specifying what filter to use on data. Typically called search condition. Defaults to empty.}

\item{debug.auth}{Logical value for getting verbose output of auth header being constructed. Defaults to false.}

\item{debug.query}{Logical value for getting verbose output of HTTP response, printing all headers. Defaults to false.}

\item{content.response}{Logical value to determine whether to retrieve full response or just the documents}

\item{max.retry}{Numeric value corresponding to number of retries on 429 error}

\item{retry.pause}{Numeric value number of seconds to pause between 429 error retrys}
}
\value{
Prints status code of HTTP POST, and returns full HTTP response or just the content
}
\description{
POST a full query to the REST API for Cosmos DB.
}
\examples{
cosmosQuery(sql.what = "c.contact.eloquaId", sql.where = "c.contact.eloquaId != null")
}
\keyword{cosmosdb}
\keyword{post}
\keyword{query}
