% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmosQuery.R
\name{cosmosQueryAll}
\alias{cosmosQueryAll}
\title{POST a full query to the REST API for Cosmos DB which returns all rows.}
\usage{
cosmosQueryAll(sql.what = "*", sql.where = "")
}
\arguments{
\item{sql.what}{String for specifying what fields to retrieve. Typically called select condition. Defaults to *}

\item{sql.where}{String for specifying what filter to use on data. Typically called search condition. Defaults to empty.}
}
\value{
Returns the content
}
\description{
POST a full query to the REST API for Cosmos DB which returns all rows.
}
\examples{
cosmosQueryAll(sql.what = "c.contact.eloquaId", sql.where = "c.contact.eloquaId != null")
}
\keyword{cosmosdb}
\keyword{post}
\keyword{query}
