% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_CARNIVAL_options.R
\name{default_CARNIVAL_options}
\alias{default_CARNIVAL_options}
\title{default_CARNIVAL_options}
\usage{
default_CARNIVAL_options()
}
\value{
returns a list with all possible options implemented in CARNIVAL.
see the documentation on \code{\link{CARNIVAL::runCARNIVAL}}.
}
\description{
generates default CARNIVAL options.
}
