% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_vector.R
\name{discretize_input}
\alias{discretize_input}
\title{discretize input vector}
\usage{
discretize_input(input_vec)
}
\arguments{
\item{input_vec}{numerical value}
}
\description{
convert a vector of continuous values to discrete {-1, 0, 1}
}
