% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_ORA.R
\name{extract_COSMOS_nodes}
\alias{extract_COSMOS_nodes}
\title{\code{extract_COSMOS_nodes}}
\usage{
extract_COSMOS_nodes(CosmosResults)
}
\arguments{
\item{CosmosResults}{COSMOS output}
}
\value{
List with 2 objects: the success and the background genes
}
\description{
Function to extract the nodes that appear in the COSMOS output network and
the background genes (all genes present in the prior knowledge network)
}
