% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleRnival.R
\name{filter_incohrent_TF_target}
\alias{filter_incohrent_TF_target}
\title{filter_incohrent_TF_target}
\usage{
filter_incohrent_TF_target(
  decouplRnival_res,
  TF_reg_net,
  meta_network,
  RNA_input
)
}
\arguments{
\item{decouplRnival_res}{A data frame resulting from the decoupleRnival function.}

\item{TF_reg_net}{A data frame containing prior knowledge of transcription factor (TF) regulatory interactions.}

\item{meta_network}{A network data frame containing signed directed prior knowledge of molecular interactions.}

\item{RNA_input}{A named vector containing differential gene expression data.}
}
\value{
A network data frame containing the genes that are not incoherently regulated by TFs.
}
\description{
Filters incoherent target genes from a regulatory network based on a decoupling analysis
of upstream and downstream gene expression.
}
\examples{

library(dplyr)

# Example input data
data("toy_RNA")
data("toy_signaling_input")
data("toy_metabolic_input")
data("toy_network")

TF_reg_net <- data.frame(
  source = c("MYC", "MYC"),
  target = c("LDHA", "SLC2A1"),
  mor = c(1, -1)
)

toy_network <- toy_network \%>\% group_by(source,target) \%>\% summarise_each(funs(mean(., na.rm = TRUE)))
toy_network <- as.data.frame(toy_network)
toy_network <- toy_network[toy_network$interaction \%in\% c(1,-1),]

# Run the decoupleRnival function to get the upstream influence scores
upstream_scores <- decoupleRnival(toy_signaling_input, toy_metabolic_input, toy_network, n_layers = 6, n_perm = 100)

# Filter incoherent target genes from the regulatory network based on the upstream influence scores
filtered_network <- filter_incohrent_TF_target(upstream_scores, TF_reg_net, toy_network, toy_RNA)

# View the resulting filtered network
print(filtered_network)
}
