% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pkn_expressed_genes.R
\name{filter_pkn_expressed_genes}
\alias{filter_pkn_expressed_genes}
\title{filter_pkn_expressed_genes}
\usage{
filter_pkn_expressed_genes(expressed_genes_entrez, meta_pkn)
}
\arguments{
\item{expressed_genes_entrez}{EntrezID of expressed genes}

\item{meta_pkn}{COSMOS prior knowledge network}
}
\description{
filters the non-expressed genes from the prior knowledge network
}
