% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_transcriptional_regulations.R
\name{filter_transcriptional_regulations}
\alias{filter_transcriptional_regulations}
\title{Filter Based On Transcriptional Regulations}
\usage{
filter_transcriptional_regulations(
  network,
  gene_expression_binarized,
  signaling_data,
  tf_regulon
)
}
\arguments{
\item{network}{Prior knowledge network (PKN).  By default COSMOS use a PKN 
derived from Omnipath, STITCHdb and Recon3D. See details on the data 
\code{\link{meta_network}}.}

\item{gene_expression_binarized}{Named vector of {-1,0,1} difnining which 
genes changed.}

\item{signaling_data}{Named vector containing known activity of signaling
nodes (kinases, TFs).}

\item{tf_regulon}{Collection of transcription factor - target interactions.
A default collection from dorothea can be obtained by the 
\code{\link{load_tf_regulon_dorothea}} function.}
}
\value{
A filtered version of the network.
}
\description{
Filters interactions between TFs and genes in the PKN by the following points:
- finds the TF target genes that change strongly (threshold)
- finds TFs from the inputs (known TF activity)
- remove interactions if the target gene expression doesn't change (regulation 
should not go through this node )
- remove interactions if the TF activity is not in agreement with the change
 in the target gene expression (other factors influence these genes)
}
