% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_COSMOS_results.R
\name{format_COSMOS_res}
\alias{format_COSMOS_res}
\title{format_COSMOS_res}
\usage{
format_COSMOS_res(cosmos_res, metab_mapping = NULL)
}
\arguments{
\item{cosmos_res}{results of COSMOS run}

\item{metab_mapping}{a named vector with HMDB Ids as names and desired metabolite names as values.}
}
\value{
list with network and attribute tables.
}
\description{
formats the network with readable names
}
