% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_COSMOS_results.R
\name{format_COSMOS_results_deprecated}
\alias{format_COSMOS_results_deprecated}
\title{graphs from COSMOS results}
\usage{
format_COSMOS_results_deprecated(
  cosmos_res,
  metab_mapping,
  gene_mapping = "ensembl",
  measured_nodes,
  omnipath_ptm
)
}
\arguments{
\item{cosmos_res}{results of CARNIVAL run}

\item{metab_mapping}{mapping table between metabolite naming conventions,
a two column dataframe with names: c("name","pubchem")}

\item{gene_mapping}{-- not sure !!!}

\item{measured_nodes}{vector of nodes that are measured or inputs}

\item{omnipath_ptm}{ptms database from OmnipathR}
}
\description{
formats the network with readable names
}
