% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesymbols_to_entrezid.R
\name{genesymbols_to_entrezid}
\alias{genesymbols_to_entrezid}
\title{convert gene symbols to entrez id}
\usage{
genesymbols_to_entrezid(symbols)
}
\arguments{
\item{symbols}{vector of genesymbols}
}
\description{
convert gene symbols to entrez id
}
