% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHeader.R
\name{genHeader}
\alias{genHeader}
\title{Generate authentication string for REST API interaction}
\usage{
genHeader(verb, resource.type = "docs", resource.link, stored.time,
  master.key = envCosmosDB, debug = FALSE)
}
\arguments{
\item{verb}{Type of operation to be done in the request}

\item{resource.type}{Target scope of the operation. Defaults to docs}

\item{resource.link}{Path to the database and collection being used}

\item{stored.time}{Time to pass in for hashing}

\item{master.key}{Access key to the database. Defaults to the envCosmosDB environment}

\item{debug}{Whether to print verbose messages or not. Defaults to FALSE}
}
\value{
String for use in the header as authentication to the CosmosDB
}
\description{
Generate authentication string for REST API interaction
}
\examples{
genHeader("POST", "docs", "dbs/dbName/colls/collName", Sys.time())
}
\keyword{authentication}
\keyword{cosmos}
\keyword{cosmosdb}
\keyword{header}
