% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TF_activity_from_CARNIVAL.R
\name{get_TF_activity_from_CARNIVAL}
\alias{get_TF_activity_from_CARNIVAL}
\title{get_TF_activity_from_CARNIVAL}
\usage{
get_TF_activity_from_CARNIVAL(carnival_result, TFs)
}
\arguments{
\item{carnival_result}{list obtained from runCARNIVAL}

\item{TFs}{character vector of transcription factors using EntrezIDs}
}
\value{
named numerical vector with TF activity found in CARNIVAL results.
}
\description{
screens the CARNIVAL results and obtains the activity of transcription factors
}
