% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_ORA.R
\name{gmt_to_dataframe}
\alias{gmt_to_dataframe}
\title{Convert gmt file to data frame}
\usage{
gmt_to_dataframe(gmtfile)
}
\arguments{
\item{gmtfile}{a full path name of the gmt file to be converted}
}
\value{
a two column data frame where the first column corresponds to omic
features and the second column to associated terms (pathways).
}
\description{
This function is designed to convert a gmt file (gene set file from MSigDB)
into a two column data frame where the first column corresponds to omic
features (genes) and the second column to associated terms (pathway the gene
belongs to). One gene can belong to several pathways.
}
