% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMDB_mapper_vec.R
\docType{data}
\name{HMDB_mapper_vec}
\alias{HMDB_mapper_vec}
\title{Toy Input Transcription Data Set}
\format{
An object of class \dQuote{\code{character}} containing the marching between HMDB metabolite IDs and there coresponding metabolite names.
}
\source{
\url{https://bioconductor.org/packages/release/data/annotation/html/metaboliteIDmapping.html}
}
\usage{
data(HMDB_mapper_vec)
}
\description{
This exemplary transcription data are the specific deregulated gene expression of the 786-O cell line from the NCI60 dataset.
}
\examples{
data(HMDB_mapper_vec)

}
\keyword{datasets}
