% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_downstream_neighbours.R
\name{keep_controllable_neighbours}
\alias{keep_controllable_neighbours}
\title{keep controllable neighbors}
\usage{
keep_controllable_neighbours(network, n_steps, input_nodes)
}
\arguments{
\item{network}{network in 3 column data.frame format
(source, interaction, target)}

\item{n_steps}{largest distance t consider}

\item{input_nodes}{names of the input nodes in the network}
}
\description{
keeps the nodes in network that are no more then n_steps away from the starting nodes.
}
