% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_downstream_neighbours.R
\name{keep_downstream_neighbours}
\alias{keep_downstream_neighbours}
\title{keep downstream neighbors}
\usage{
keep_downstream_neighbours(network, n_steps, starting_nodes)
}
\arguments{
\item{network}{network in 3 column data.frame format
(source, interaction, target)}

\item{n_steps}{largest distance t consider}

\item{starting_nodes}{names of nodes in the network}
}
\description{
keeps the nodes in network that are no more then n_steps away from the starting nodes.
}
