% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_meta_pkn.R
\name{load_meta_pkn}
\alias{load_meta_pkn}
\title{load prior knowledge network}
\usage{
load_meta_pkn()
}
\description{
prior knowledge network derived from Omnipath, STITCHdb and Recon3D.
See details in the paper.
}
\details{
TODO: describe what are the names used in the PKN and how the PKN was constructed.
}
