% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tf_regulon_dorothea_omnipath.R
\name{load_tf_regulon_dorothea_omnipath}
\alias{load_tf_regulon_dorothea_omnipath}
\title{load_tf_regulon_dorothea_omnipath}
\usage{
load_tf_regulon_dorothea_omnipath()
}
\description{
downloads the TF-regulons from Omnipaht. Different from DOROTHEA regulon because
sign is handled differently: if both stimulation and inhibition was reported then
it is removed.
}
