% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metab_to_pubchem.R
\docType{data}
\name{metab_to_pubchem}
\alias{metab_to_pubchem}
\title{mapping between metabolite names and pubchem CIDs}
\format{
An object of class \code{"data.frame"};
}
\usage{
data(metab_to_pubchem)
}
\description{
mapping between metabolite names and pubchem CIDs
}
\examples{
data(metab_to_pubchem)
View(metab_to_pubchem)
}
\keyword{mapping}
