% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmos_data.R
\name{new_cosmos_data}
\alias{new_cosmos_data}
\title{create new cosmos data}
\usage{
new_cosmos_data(
  meta_network = data.frame(),
  tf_regulon = data.frame(),
  signaling_data = double(),
  metabolic_data = double(),
  expression_data = double()
)
}
\description{
Constructor. Should not be exported. The object is not validated.
}
