% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_vectors.R
\name{prepare_metab_data}
\alias{prepare_metab_data}
\title{format COSMOS metabolic input}
\usage{
prepare_metab_data(metab_data, pkn, format_names = TRUE)
}
\arguments{
\item{metab_data}{A named numeric vector, containing the values to be used for the metabolic layer in COSMOS.}

\item{pkn}{Prior knowledge network created with \code{\link{load_meta_pkn()}}.}

\item{format_names}{Add "XMetab" to the vector names?}
}
\value{
A new vector ready to be used as COSMOS input.
}
\description{
format COSMOS metabolic input
}
