% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_metab_inputs.R
\name{prepare_metab_inputs}
\alias{prepare_metab_inputs}
\title{add metabolic compartment and metab__ prefix to metabolite IDs}
\usage{
prepare_metab_inputs(metab_input, compartment_codes)
}
\arguments{
\item{metab_input}{a named  vector with matebolic statistics as inputs and metabolite identifiers as names}

\item{compartment_codes}{a character vector, the desired compartment codes to be added. Possible values are "r", "c", "e", "x", "m", "l", "n" and "g"}
}
\value{
a named vector with the compartment code and prefixed added to the names
}
\description{
This function adds metabolic compartments to the metabolic identifiers provided by the user.
}
