% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_vectors.R
\name{prepare_signaling_data}
\alias{prepare_signaling_data}
\title{format COSMOS signaling input}
\usage{
prepare_signaling_data(signaling_data, pkn, format_names = TRUE)
}
\arguments{
\item{signaling_data}{A named numeric vector, containing the values to be used for the signaling layer in COSMOS.}

\item{pkn}{Prior knowledge network created with \code{\link{load_meta_pkn()}}.}

\item{format_names}{Add "X" to the vector names?}
}
\value{
A new vector ready to be used as COSMOS input.
}
\description{
format COSMOS signaling input
}
