% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_COSMOS.R
\name{preprocess_COSMOS}
\alias{preprocess_COSMOS}
\title{preprocess_COSMOS}
\usage{
preprocess_COSMOS(
  meta_network = load_meta_pkn(),
  tf_regulon = load_tf_regulon_dorothea(),
  signaling_data,
  metabolic_data,
  expression_data,
  filter_tf_gene_interaction_by_optimization = TRUE,
  solver_path = NULL,
  solver = "cplex",
  time_limit = 3600
)
}
\arguments{
\item{meta_network}{prior knowledge network. By default COSMOS use a PKN 
derived from Omnipath, STITCHdb and Recon3D. See details on the function 
\cite{\code{load_meta_pkn}}}

\item{signaling_data}{numerical vector, where names are signaling nodes 
in the PKN and values are from \{1, 0, -1\}. Continuous data will be 
discretized using the \link\code{sign} function.}

\item{metabolic_data}{numerical vector, where names are metabolic nodes 
in the PKN and values are continuous values. These values are compared to 
with the the simulation [? range of values]}

\item{expression_data}{(optional) numerical vector, where names are gene
names and values are from \{1,-1\}}

\item{filter_tf_gene_interaction_by_optimization}{(default:TRUE), if TRUE then runs 
a network optimization that estimates TF activity not included in the inputs
and checks the consistency between the estimated activity and change in gene 
expression. Removes interactions where TF and gene expression are inconsistent}

\item{solver_path}{argument passed to \link\code{CARNIVAL::runCARNIVAL}. 
used if filter_tf_gene_interaction_by_optimization is TRUE}

\item{solver}{argument passed to \link\code{CARNIVAL::runCARNIVAL}
used if filter_tf_gene_interaction_by_optimization is TRUE}

\item{time_limit}{argument passed to \link\code{CARNIVAL::runCARNIVAL}
used if filter_tf_gene_interaction_by_optimization is TRUE}
}
\value{
named list with the following members: 
 - `meta_network`  filtered PKN
 - `tf_regulon`  TF - target regulatory network
 - `signaling_data_bin` binarised signaling data 
 - `metabolic_data`  metabolomics data
 - `expression_data_bin`  binarised gene expression data 
 - `optimized_network` initial optimized network if filter_tf_gene_interaction_by_optimization is TRUE.
}
\description{
runs checks on the input data and simplifies the prior knowledge network.
Simplification includes the removal of (1) nodes that are not reachable from 
signaling nodes and  (2) interactions between transcription factors and target
genes if the target gene does not respond or the response is contradictory
with the change in the transcription factor activity. 
Optionally, further TF activities are estimated via network optimization and 
the interactions between TF and genes are filtered again.
}
