% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_CARNIVAL_results.R
\name{process_CARNIVAL_results}
\alias{process_CARNIVAL_results}
\title{process_CARNIVAL_results}
\usage{
process_CARNIVAL_results(CARNIVAL_results)
}
\arguments{
\item{CARNIVAL_results}{list of matrices received from 
CRANIVAL::run_CARNIVAL()}
}
\value{
list with the following elements:
- `aggregated_network` the averaged networks found by optimization in a 
format of a Simple Interaction network, i.e. each row codes an edge
- `N_networks`: number of solutions found by the optimization
- `aggregated_network_node_attributes`: estimated node properties
- `individual_networks`: list of optimial networks found
- `individual_networks_node_attributes`: node activity in each network.
}
\description{
formats the raw CARNIVAL results to a more appealing list of networks.
}
