% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readContent.R
\name{readContent}
\alias{readContent}
\title{Decode the raw contents of an HTTP response and convert it to a list}
\usage{
readContent(binary.stream)
}
\arguments{
\item{binary.stream}{The content field of an HTTP response to be decoded as a character}
}
\value{
List variable of the 'contents' field from HTTP response
}
\description{
Decode the raw contents of an HTTP response and convert it to a list
}
\examples{
readContent(http.response)
}
\keyword{binary}
\keyword{character}
\keyword{json}
\keyword{raw}
\keyword{response}
