% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_COSMOS.R
\name{run_COSMOS_signaling2metabolism}
\alias{run_COSMOS_signaling2metabolism}
\title{run_COSMOS_s2m}
\usage{
run_COSMOS_signaling2metabolism(
  meta_network = load_meta_pkn(),
  tf_regulon = load_tf_regulon_dorothea(),
  signaling_data,
  metabolic_data,
  expression_data,
  solver_path = NULL,
  solver = "cplex",
  time_limit = 3600,
  test_run = FALSE
)
}
\arguments{
\item{meta_network}{prior knowledge network. By default COSMOS use a PKN 
derived from Omnipath, STITCHdb and Recon3D. See details on the function 
\cite{\code{load_meta_pkn}}}

\item{signaling_data}{numerical vector, where names are signaling nodes 
in the PKN and values are from \{1, 0, -1\}. Continuous data will be 
discretized using the \link\code{sign} function.}

\item{metabolic_data}{numerical vector, where names are metabolic nodes 
in the PKN and values are continuous values. These values are compared to 
with the the simulation [? range of values]}

\item{expression_data}{(optional) numerical vector, where names are gene
names and values are from \{1,-1\}}

\item{solver_path}{argument passed to \link\code{CARNIVAL::runCARNIVAL}}

\item{solver}{argument passed to \link\code{CARNIVAL::runCARNIVAL}}

\item{time_limit}{argument passed to \link\code{CARNIVAL::runCARNIVAL}}

\item{test_run}{(default:FALSE), if TRUE then runs the pipeline without the
last optimization step. Useful for diagnostics.}
}
\description{
runs COSMOS from signaling to metabolism
}
