% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_ORA.R
\name{top_pathways}
\alias{top_pathways}
\title{Plot top ORA pathways over all sources}
\usage{
top_pathways(sigPathwaysDf)
}
\arguments{
\item{sigPathwaysDf}{Data frame containing all significant pathways as a result of the ORA}
}
\value{
Bar plot of the 20 most significant pathways by p-value
}
\description{
Function to plot the most significantly over-expressed pathways (max. 20)
(regardless of source, e.g. Reactome, Biocarta)
}
\examples{
top_pathways()
}
