% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_metabolic_input.R
\docType{data}
\name{toy_metabolic_input}
\alias{toy_metabolic_input}
\title{Toy Metabolic Input Data}
\format{
An object of class \dQuote{\code{numeric}} containing the t-values of
  2 metabolites, which are named with metabolite HMDB Ids matching the
  toy network.
}
\source{
Subset of: 
  \url{https://github.com/saezlab/COSMOS_MSB/blob/main/data/metab_input_COSMOS.csv}
}
\usage{
data(toy_metabolic_input)
}
\description{
This metabolic data are a subset from the metabolic measurements of the 786-O cell line from the NCI60 dataset.
}
\examples{
data(toy_metabolic_input)

}
\references{
{
  Dugourd, A., Kuppe, C. and Sciacovelli, M. et. al. (2021) \emph{Molecular 
  Systems Biology}. \bold{17}, e9730.
}
}
\keyword{datasets}
