% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_COSMOS_inputs.R
\name{check_COSMOS_inputs}
\alias{check_COSMOS_inputs}
\title{check_COSMOS_inputs}
\usage{
check_COSMOS_inputs(
  meta_network = NULL,
  tf_regulon = NULL,
  signaling_data = NULL,
  metabolic_data = NULL,
  expression_data = NULL
)
}
\arguments{
\item{meta_network}{prior knowledge network}

\item{tf_regulon}{transcription factor regulon}

\item{signaling_data}{named numerical vector}

\item{metabolic_data}{named numerical vector}

\item{expression_data}{named numerical vector}
}
\description{
checks the format of the main data inputs. Checks only the data that was 
passed (not NULL).
}
