% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs_CARNIVAL.R
\name{check_inputs_for_CARNIVAL}
\alias{check_inputs_for_CARNIVAL}
\title{check_inputs_for_CARNIVAL}
\usage{
check_inputs_for_CARNIVAL(meta_network, input_data, measured_data)
}
\description{
checks the format of the main data inputs for CARNIVL. Checks the data format
and coverage of nodes in the PKN and data. All nodes in input_data and 
measured data must apper in the PKN
}
