% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ensembl_to_entrezid.R
\name{convert_ensembl_to_entrezid}
\alias{convert_ensembl_to_entrezid}
\title{convert gene ensembl to entrez id}
\usage{
convert_ensembl_to_entrezid(ensembl)
}
\arguments{
\item{ensembl}{vector of genes with ensembl id}
}
\description{
convert gene ensembl to entrez id
}
\seealso{
[convert_genesymbols_to_entrezid()]
}
