% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_genesymbols_to_entrezid.R
\name{convert_genesymbols_to_entrezid}
\alias{convert_genesymbols_to_entrezid}
\title{convert gene symbols to entrez id}
\usage{
convert_genesymbols_to_entrezid(symbols)
}
\arguments{
\item{symbols}{vector of genesymbols}
}
\description{
convert gene symbols to entrez id
}
\seealso{
[convert_ensembl_to_entrezid()]
}
