% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmos_data.R
\name{cosmos_data}
\alias{cosmos_data}
\title{create cosmos data}
\usage{
cosmos_data(
  meta_network,
  tf_regulon = NULL,
  signaling_data,
  metabolic_data,
  expression_data,
  verbose = TRUE
)
}
\description{
user-friendly interface to create cosmos data.
Based on the check_cosmos_inputs, but this function removes the non-overlapping
nodes rather than reporting error.
}
