% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmos_data.R
\name{cosmos_data}
\alias{cosmos_data}
\title{create cosmos data}
\usage{
cosmos_data(
  meta_network,
  tf_regulon = NULL,
  signaling_data,
  metabolic_data,
  expression_data,
  verbose = TRUE
)
}
\description{
An S3 class that combines the required data into a comprehensive list. 
Use the \code{\link{preprocess_COSMOS_signaling_to_metabolism}} or 
\code{\link{preprocess_COSMOS_metabolism_to_signaling}} to create an instance.
}
