% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleRnival.R
\name{decoupleRnival}
\alias{decoupleRnival}
\title{DecoupleRnival}
\usage{
decoupleRnival(
  upstream_input = NULL,
  downstream_input,
  meta_network,
  n_layers,
  n_perm = 1000
)
}
\arguments{
\item{upstream_input}{A named vector with up_stream nodes and their corresponding activity.}

\item{downstream_input}{A named vector with down_stream nodes and their corresponding activity.}

\item{meta_network}{A network data frame containing signed directed prior knowledge of molecular interactions.}

\item{n_layers}{The number of layers that will be propagated upstream.}

\item{n_perm}{The number of permutations to use in the decoupling algorithm.}
}
\value{
A data frame containing the score of the nodes upstream of the 
downstream input based on the itterative propagation
}
\description{
Itterativelly propagate downstream input activity through a signed directed 
using the weighted mean of decoupleR
}
