% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_CARNIVAL_options.R
\name{default_CARNIVAL_options}
\alias{default_CARNIVAL_options}
\title{Setting Default CARNIVAL Options}
\usage{
default_CARNIVAL_options()
}
\value{
returns a list with all possible options implemented in CARNIVAL.
see the documentation on \code{\link[CARNIVAL]{runCARNIVAL}}.
}
\description{
Returns the default CARNIVAL options as a list.  You can modify the elements
of the list and then use it as an argument in \dQuote{\code{run_COSMOS()}}.
}
\examples{
# load and change default options: 
my_options = default_CARNIVAL_options()

my_options$solverPath = "/Applications/CPLEX_Studio128/cplex/bin/x86-64_osx/cplex"
my_options$threads = 2
my_options$timelimit = 3600*15
}
