% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleRnival.R
\name{filter_incohrent_TF_target}
\alias{filter_incohrent_TF_target}
\title{filter_incoherent_TF_targets}
\usage{
filter_incohrent_TF_target(
  decouplRnival_res,
  TF_reg_net,
  meta_network,
  RNA_input
)
}
\arguments{
\item{decouplRnival_res}{A data frame resulting from the decoupleRnival function.}

\item{TF_reg_net}{A data frame containing prior knowledge of transcription factor (TF) regulatory interactions.}

\item{meta_network}{A network data frame containing signed directed prior knowledge of molecular interactions.}

\item{RNA_input}{A named vector containing differential gene expression data.}
}
\value{
A network data frame containing the genes that are not incoherently regulated by TFs.
}
\description{
Filters incoherent target genes from a regulatory network based on a decoupling analysis
of upstream and downstream gene expression.
}
