% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_transcriptional_regulations.R
\name{filter_transcriptional_regulations}
\alias{filter_transcriptional_regulations}
\title{filter_transcriptional_regulations}
\usage{
filter_transcriptional_regulations(
  network,
  gene_expression_binarized,
  signaling_data,
  tf_regulon
)
}
\description{
Filters interactions between TFs and genes in the PKN by the following points
- finds the TF target genes that change strongly (threshold)
- finds TFs from the inputs (known TF activity)
- remove interactions if the target gene expression doesn't change (regulation 
should not go through this node )
- remove interactions if the TF activity is not in agreement with the change
 in the target gene expression (other factors influence these genes)
 
 @param network meta PKN
 @param gene_expression_binarized named vector of {-1,0,1} difnining which genes changed
 @param signaling_data named vector containing known activity of signaling nodes (kinases, TFs)
 @param tf_regulon transcription factor regulon, e.g. from dorothea
 @return a filtered version of the network
 @importFrom rlang .data
}
