% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tf_regulon_dorothea.R
\name{load_tf_regulon_dorothea}
\alias{load_tf_regulon_dorothea}
\title{load transcription factor regulon}
\usage{
load_tf_regulon_dorothea(toEntrez = TRUE, confidence = c("A", "B", "C"))
}
\arguments{
\item{toEntrez}{if TRUE (default), converts gene symbols to EntrezID}
}
\description{
load the transcription factors from \code{DOROTHEA} package and converts 
gene symbols to EntrezID using biomart
}
