% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tf_regulon_dorothea.R
\name{load_tf_regulon_dorothea}
\alias{load_tf_regulon_dorothea}
\title{load transcription factor regulon}
\usage{
load_tf_regulon_dorothea(toEntrez = TRUE, confidence = c("A", "B", "C"))
}
\arguments{
\item{toEntrez}{if TRUE (default), converts gene symbols to EntrezID}

\item{confidence}{strong vector (by default: c("A","B","C")). Subset of \{A, B,
 C, D, E\}. See the `dorothea` for the meaning of confidence levels. 
package for further details.}
}
\value{
returns a PKN of a form of a data table. Each row is an interaction.
Columns names are:
- `tf` transcription factor
- `confidence` class of confidence
- `target` target gene
- `sign` indicates if interaction is up (1) or down-regulation (-1).
}
\description{
load the transcription factors from \code{DOROTHEA} package and converts 
gene symbols to EntrezID using biomart
}
