% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_ORA.R
\name{plot_pathways}
\alias{plot_pathways}
\title{\code{plot_pathways}}
\usage{
plot_pathways(sigPathwaysDf, cutoff)
}
\arguments{
\item{sigPathwaysDf}{Data frame containing all significant pathways as a result of the ORA}

\item{cutoff}{number used as a filter criteria for p-value}
}
\value{
plot of the 5 most significant pathways per source
}
\description{
Function to display the most significantly over-expressed pathways sorted by
their source (e.g. Reactome, Biocarta, etc.)
}
