% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_vectors.R
\name{prepare_metab_data}
\alias{prepare_metab_data}
\title{format COSMOS metabolic input}
\usage{
prepare_metab_data(metabolic_data, meta_network)
}
\arguments{
\item{metabolic_data}{A named numeric vector, containing the values to be used for the metabolic layer in COSMOS. The names of the vector should be PUBCHEM IDs.}

\item{meta_network}{Prior knowledge network created with \code{\link{load_meta_pkn()}}.}
}
\value{
A new vector ready to be used as COSMOS input.
}
\description{
This function prepares the metabolic data to be used in the COSMOS optimization steps. It takes as input a vector with the metabolic data (e.g, limma t values)
named with PUBCHEM IDs and expand it to the multi-compartment COSMOS format. It also messages the number of final inputs in the meta network.
}
