% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_CARNIVAL_results.R
\name{process_CARNIVAL_results}
\alias{process_CARNIVAL_results}
\title{process_CARNIVAL_results}
\usage{
process_CARNIVAL_results(CARNIVAL_results)
}
\arguments{
\item{CARNIVAL_results}{list of matrices received from 
CRANIVAL::run_CARNIVAL()}
}
\value{
list with the following elements:
  \describe{
    \item{\code{aggregated_network}}{The averaged networks found by 
    optimization in a format of a Simple Interaction network, i.e. each row 
    codes an edge}
    \item{\code{N_networks}}{Number of solutions found by the 
    optimization}
    \item{\code{aggregated_network_node_attributes}}{Estimated node properties}
    \item{\code{individual_networks}}{List of optimial networks found}
    \item{\code{individual_networks_node_attributes}}{Node activity in each
    network}
  }
}
\description{
formats the raw CARNIVAL results to a more appealing list of networks.
}
