% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoupleRnival.R
\name{reduce_solution_network}
\alias{reduce_solution_network}
\title{reduce_solution_network}
\usage{
reduce_solution_network(
  decoupleRnival_res,
  meta_network,
  cutoff,
  upstream_input,
  RNA_input = NULL,
  n_steps = 10
)
}
\arguments{
\item{decoupleRnival_res}{A data frame resulting from the decoupleRnival function.}

\item{meta_network}{A network data frame containing signed directed prior knowledge of molecular interactions.}

\item{cutoff}{The consistency threshold for filtering edges from the solution network.}

\item{upstream_input}{A named vector with up_stream nodes and their corresponding activity.}

\item{RNA_input}{A named vector containing differential gene expression data.}

\item{n_steps}{The maximum number of steps from upstream input nodes to include in the solution network.}
}
\value{
A list containing the solution network (SIF) and an attribute table (ATT) with gene expression data.
}
\description{
Reduces a solution network based on a decoupling analysis of upstream and downstream gene expression,
by filtering out edges that do not meet a consistency threshold, and limiting the network to a
certain number of steps from upstream input nodes.
}
