% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_COSMOS_metabolism_to_signaling.R
\name{run_COSMOS_metabolism_to_signaling}
\alias{run_COSMOS_metabolism_to_signaling}
\title{COSMOS metabolism to signaling}
\usage{
run_COSMOS_metabolism_to_signaling(
  meta_network,
  signaling_data,
  metabolic_data,
  CARNIVAL_options = default_CARNIVAL_options(),
  test_run = FALSE
)
}
\arguments{
\item{meta_network}{prior knowledge network.}

\item{signaling_data}{numerical vector, where names are signaling nodes 
in the PKN and values are continuous values that represents log2 fold change 
or t values from a differential analysis. These values are compared to 
the simulation results (simulated nodes can take value -1, 0 or 1)}

\item{metabolic_data}{numerical vector, where names are metabolic nodes 
in the PKN and values are from \{1, 0, -1\}. Continuous data will be 
discretized using the \code{\link{sign}} function.}

\item{CARNIVAL_options}{list that controls the options of CARNIVAL. See details 
in \code{\link{default_CARNIVAL_options()}}.}
}
\value{
named list with the following members: 
 - `meta_network`  filtered PKN
 - `tf_regulon`  TF - target regulatory network
 - `signaling_data_bin` binarised signaling data 
 - `metabolic_data`  metabolomics data
 - `optimized_network` initial optimized network if filter_tf_gene_interaction_by_optimization is TRUE.
}
\description{
runs COSMOS from metabolism to signaling. This function finds a subset of the
prior knowledge network based on optimisation that (1) includes the most 
measured and input nodes and (2) which is in agreement with the data. 
Use \code{\link{preprocess_COSMOS}} to prepare the prior knowledge network or 
load the one in the toolbox.
}
