% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_COSMOS_signaling_to_metabolism.R
\name{run_COSMOS_signaling_to_metabolism}
\alias{run_COSMOS_signaling_to_metabolism}
\title{run_COSMOS_signaling2metabolism}
\usage{
run_COSMOS_signaling_to_metabolism(
  meta_network,
  signaling_data,
  metabolic_data,
  solver_path = NULL,
  solver = "cplex",
  time_limit = 7200,
  test_run = FALSE
)
}
\arguments{
\item{meta_network}{prior knowledge network.}

\item{signaling_data}{numerical vector, where names are signaling nodes 
in the PKN and values are from \{1, 0, -1\}. Continuous data will be 
discretized using the \link{\code{sign}} function.}

\item{metabolic_data}{numerical vector, where names are metabolic nodes 
in the PKN and values are continuous values that represents log2 fold change 
or t values from a differential analysis. These values are compared to 
the simulation results (simulated nodes can take value -1, 0 or 1)}

\item{solver_path}{argument passed to \link{\code{CARNIVAL::runCARNIVAL}}. 
used if filter_tf_gene_interaction_by_optimization is TRUE}

\item{solver}{argument passed to \link{\code{CARNIVAL::runCARNIVAL}}
used if filter_tf_gene_interaction_by_optimization is TRUE}

\item{time_limit}{argument passed to \link{\code{CARNIVAL::runCARNIVAL}}
used if filter_tf_gene_interaction_by_optimization is TRUE}
}
\value{
named list with the following members: 
 - `meta_network`  filtered PKN
 - `tf_regulon`  TF - target regulatory network
 - `signaling_data_bin` binarised signaling data 
 - `metabolic_data`  metabolomics data
 - `optimized_network` initial optimized network if filter_tf_gene_interaction_by_optimization is TRUE.
}
\description{
runs COSMOS from signaling to metabolism. This function finds a subset of the
prior knowledge network based on optimisation that (1) includes the most 
measured and input nodes and (2) which is in agreement with the data. 
Use \link{\code{preprocess_COSMOS}} to prepare the prior knowledge network or 
load the one in the toolbox.
}
