% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_COSMOS_signaling_to_metabolism.R
\name{run_COSMOS_signaling_to_metabolism}
\alias{run_COSMOS_signaling_to_metabolism}
\title{run COSMOS signaling to metabolism}
\usage{
run_COSMOS_signaling_to_metabolism(
  data,
  CARNIVAL_options = default_CARNIVAL_options(),
  test_run = FALSE
)
}
\arguments{
\item{data}{\code{\link{cosmos_data}} object. Use the \code{\link{preprocess_COSMOS_signaling_to_metabolism}}
function to create one.}

\item{CARNIVAL_options}{list that controls the options of CARNIVAL. See details 
in \code{\link{default_CARNIVAL_options}}.}
}
\value{
list with the following elements:
- `aggregated_network` the averaged networks found by optimization in a 
format of a Simple Interaction network, i.e. each row codes an edge
- `N_networks`: number of solutions found by the optimization
- `aggregated_network_node_attributes`: estimated node properties
- `individual_networks`: list of optimial networks found
- `individual_networks_node_attributes`: node activity in each network.
}
\description{
runs COSMOS from signaling to metabolism. This function finds a subset of the
prior knowledge network based on optimisation that (1) includes the most 
measured and input nodes and (2) which is in agreement with the data. 
Use \code{\link{preprocess_COSMOS_signaling_to_metabolism}} to prepare 
inputs, measurements and prior knowledge network.
}
