% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCARNIVAL_wrapper.R
\name{runCARNIVAL_wrapper}
\alias{runCARNIVAL_wrapper}
\title{runCARNIVAL_wrapper}
\usage{
runCARNIVAL_wrapper(
  network,
  input_data,
  measured_data,
  solver_path,
  solver,
  timelimit,
  mipGAP
)
}
\arguments{
\item{network}{prior knowledge network. data.frame object with source, sign and 
target columns}

\item{input_data}{numerical vector, where names are input nodes 
in the PKN and values are from \{1, 0, -1\}.}

\item{measured_data}{numerical vector, where names are measured nodes 
in the PKN and values are continuous values. These values are compared to 
with the the simulation}

\item{solver_path}{argument passed to \link{\code{CARNIVAL::runCARNIVAL}}}

\item{solver}{argument passed to \link{\code{CARNIVAL::runCARNIVAL}}}

\item{time_limit}{argument passed to \link{\code{CARNIVAL::runCARNIVAL}}}
}
\description{
checks and formats the COSMOS data to CARNIVAL inputs and runs CARNIVAL.
}
