% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_metabolic_input.R
\docType{data}
\name{toy_metabolic_input}
\alias{toy_metabolic_input}
\title{Toy Metabolic Input Data}
\format{
An object of class \dQuote{\code{numeric}} containing the t-values of
  3 metabolites, which are named with metabolite PubChem CIDs matching the
  toy network.
}
\source{
Subset of: 
  \url{https://github.com/saezlab/COSMOS_MSB/blob/main/data/metab_input_COSMOS.csv}
}
\usage{
data(toy_metabolic_input)
}
\description{
This metabolic data are a subset from the metabolic measurements used as an
input in the case study of the COSMOS paper.  The subset contains a random 
selection of metabolites present in the toy network.
}
\examples{
toy_metabolic_input = toy_metabolic_input

}
\references{
{
  Dugourd, A., Kuppe, C. and Sciacovelli, M. et. al. (2021) \emph{Molecular 
  Systems Biology}. \bold{17}, e9730.
}
}
\keyword{datasets}
