% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TF_activity_from_CARNIVAL.R
\name{validate_CARNIVAL_results}
\alias{validate_CARNIVAL_results}
\title{Validate CARNIVAL Results}
\usage{
validate_CARNIVAL_results(CR)
}
\arguments{
\item{CR}{CARNIVAL result object}
}
\description{
Implement here basic checks to see if we got back from CARNIVAL what we
expected.  Subject to change if CARNIVAL results changes.
}
